GCP/pipeline-ecommerce-bq-dbt/cloud_run_dbt/requirements.txt
dbt-bigquery==1.9.1
Flask==3.0.3
google-cloud-logging==3.11.4